feat : 마이페이지 API 연결

See merge request s11-webmobile1-sub2/S11P12A107!122

+++ backend/src/main/java/com/ssafy/hifive/domain/reservation/service/ReservationService.java
@@ -6,7 +6,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.ssafy.hifive.domain.fanmeeting.entity.Fanmeeting;
 import com.ssafy.hifive.domain.fanmeeting.repository.FanmeetingRepository;
 import com.ssafy.hifive.domain.member.entity.Member;
-import com.ssafy.hifive.domain.reservation.handler.ReservationWebSocketHandler;
+import com.ssafy.hifive.domain.reservation.dto.response.ReservationMemberDto;
 import com.ssafy.hifive.global.error.ErrorCode;
 import com.ssafy.hifive.global.error.type.BadRequestException;
 import com.ssafy.hifive.global.error.type.DataNotFoundException;
@@ -18,24 +18,19 @@ import lombok.RequiredArgsConstructor;
 public class ReservationService {
 	private final FanmeetingRepository fanmeetingRepository;
 	private final ReservationFanmeetingPayService reservationFanmeetingPayService;
-	private final ReservationWebSocketHandler reservationWebSocketHandler;
 	private final ReservationQueueService reservationQueueService;
 	private final ReservationFanmeetingReserveService reservationFanmeetingReserveService;
 	private final ReservationValidService reservationValidService;
 
-	public void reserve(long fanmeetingId, Member member) {
+	public ReservationMemberDto reserve(long fanmeetingId, Member member) {
 		Fanmeeting fanmeeting = fanmeetingRepository.findById(fanmeetingId)
 			.orElseThrow(() -> new DataNotFoundException(ErrorCode.FANMEETING_NOT_FOUND));
 
 		reservationFanmeetingReserveService.checkReservation(fanmeeting, member);
 
-		String queueKey = "fanmeeting:" + fanmeetingId + ":waiting-queue";
-		String payingQueueKey = "fanmeeting:" + fanmeetingId + ":paying-queue";
-		if(reservationValidService.addToPayingQueueIsValid(payingQueueKey)){
-			reservationQueueService.addToPayingQueue(payingQueueKey, member.getMemberId());
-		} else {
-			reservationQueueService.addToWaitingQueue(queueKey, member.getMemberId());
-		}
+		addToQueue(fanmeetingId, member.getMemberId());
+
+		return ReservationMemberDto.from(member);
 	}
 
 	@Transactional
@@ -43,23 +38,22 @@ public class ReservationService {
 		Fanmeeting fanmeeting = fanmeetingRepository.findById(fanmeetingId)
 			.orElseThrow(() -> new DataNotFoundException(ErrorCode.FANMEETING_NOT_FOUND));
 
+		String payingQueueKey = "fanmeeting:" + fanmeetingId + ":paying-queue";
+		if (reservationValidService.isPaymentSessionExpired(payingQueueKey, member.getMemberId())) {
+			checkAndMoveQueues(fanmeetingId);
+			throw new BadRequestException(ErrorCode.PAYMENT_SESSION_EXPIRED);
+		}
+
 		int remainingTicket = reservationFanmeetingPayService.checkRemainingTicket(fanmeeting);
 
 		reservationFanmeetingPayService.payTicket(fanmeeting, member, remainingTicket);
 
 		reservationFanmeetingPayService.recordReservation(fanmeeting, member);
 
-		String payingQueueKey = "fanmeeting:" + fanmeetingId + ":paying-queue";
-		try {
-			reservationQueueService.removeFromPayingQueue(payingQueueKey, member.getMemberId());
-			checkAndMoveQueues(fanmeetingId);
-
-		} catch (Exception e) {
-			throw new BadRequestException(ErrorCode.WEBSOCKET_MESSAGE_SEND_ERROR);
-		}
+		reservationQueueService.removeFromPayingQueue(payingQueueKey, member.getMemberId());
+		checkAndMoveQueues(fanmeetingId);
 	}
 
-	//이걸 스케줄링으로 처리할지, 지금처럼 한 명 나가면 체크하는 로직으로 짤지 고민 중
 	private void checkAndMoveQueues(long fanmeetingId) {
 		String waitingQueueKey = "fanmeeting:" + fanmeetingId + ":waiting-queue";
 		String payingQueueKey = "fanmeeting:" + fanmeetingId + ":paying-queue";
@@ -69,11 +63,22 @@ public class ReservationService {
 
 		if (slotsAvailable > 0) {
 			try {
-				reservationQueueService.moveFromWaitingToPayingQueue(waitingQueueKey, payingQueueKey, slotsAvailable);
+				reservationQueueService.moveFromWaitingToPayingQueue(fanmeetingId, waitingQueueKey, payingQueueKey,
+					slotsAvailable);
 			} catch (Exception e) {
 				throw new BadRequestException(ErrorCode.WEBSOCKET_MESSAGE_SEND_ERROR);
 			}
 
+		}
+	}
+
+	public void addToQueue(Long fanmeetingId, Long memberId) {
+		String queueKey = "fanmeeting:" + fanmeetingId + ":waiting-queue";
+		String payingQueueKey = "fanmeeting:" + fanmeetingId + ":paying-queue";
+		if (reservationValidService.addToPayingQueueIsValid(payingQueueKey)) {
+			reservationQueueService.addToPayingQueue(payingQueueKey, memberId, fanmeetingId);
+		} else {
+			reservationQueueService.addToWaitingQueue(queueKey, memberId);
 		}
 	}
 }
